+++ src/contentscript/overlay/root/components/InputPanelSettings/InputPanelSettings.module.scss
@@ -1,6 +1,7 @@
 @import "../../assets/styles/null.scss";
 @import "../../assets/styles/fonts.scss";
 @import "../../assets/styles/color.variables.scss";
+@import "../../assets/styles/animate.scss";
 
 .formDefault {
   display: flex;
@@ -68,97 +69,4 @@
   font-weight: 400;
 
   animation-fill-mode: forwards;
-}
-
-.formAbsolute {
-}
-
-@keyframes openError {
-  0% {
-    margin: 0;
-    padding: 0;
-
-    opacity: 0;
-
-    font-size: 2px;
-
-    line-height: 0%;
-  }
-
-  3% {
-    margin: 0 0 -5px;
-    padding: 2.5px 7.5px;
-
-    opacity: 0;
-
-    font-size: 7px;
-    line-height: 20%;
-  }
-
-  7% {
-    margin: 0 0 -10px;
-    padding: 5px 13px;
-
-    opacity: 0;
-  }
-
-  10% {
-    margin: 0 0 -10px;
-    padding: 5px 13px;
-
-    opacity: 1;
-
-    font-size: 14px;
-    line-height: 100%;
-  }
-
-  85% {
-    margin: 0 0 -10px;
-    padding: 5px 13px;
-
-    opacity: 1;
-
-    font-size: 14px;
-    line-height: 100%;
-  }
-
-  87% {
-    margin: 0 0 -7.5px;
-    padding: 4px 7.5px;
-
-    opacity: 0;
-
-    font-size: 12px;
-    line-height: 50%;
-  }
-
-  90% {
-    margin: 0 0 -5px;
-    padding: 2.5px 7.5px;
-
-    opacity: 0;
-
-    font-size: 7px;
-    line-height: 20%;
-  }
-
-  95% {
-    margin: 0 0 -2.5px;
-    padding: 1.5px 5px;
-
-    opacity: 0;
-
-    font-size: 2px;
-    line-height: 0%;
-  }
-
-  100% {
-    margin: 0;
-    padding: 0;
-
-    opacity: 0;
-
-    font-size: 0;
-    line-height: 0%;
-  }
 }
