+++ src/contentscript/overlay/root/components/Localhost/adaptive.scss
@@ -1,0 +1,23 @@
+@import "../../assets/styles/null.scss";
+@import "../../assets/styles/fonts.scss";
+@import "../../assets/styles/color.variables.scss";
+
+@media (min-width: 320px) and (max-width: 640px) {
+  .label {
+    display: inline-block;
+    overflow: hidden;
+
+    width: 100%;
+    max-width: 100%;
+  }
+
+  .buttonLocalhost {
+    margin-right: 0;
+  }
+
+  .labelLocalhost {
+    flex: 1 1 auto;
+
+    max-width: 65%;
+  }
+}
