+++ src/contentscript/overlay/root/components/Modal/Modal.module.scss
@@ -1,6 +1,7 @@
 @import "../../assets/styles/null.scss";
 @import "../../assets/styles/fonts.scss";
 @import "../../assets/styles/color.variables.scss";
+@import "../../assets/styles/animate.scss";
 
 .modal {
   position: fixed;
@@ -99,24 +100,4 @@
 
   font-weight: 700;
   line-height: 26px;
-}
-
-@keyframes appear {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes slide-in {
-  from {
-    transform: translateY(-150px);
-  }
-
-  to {
-    transform: translateY(0);
-  }
 }
