+++ src/contentscript/overlay/root/components/Notification/adaptive.scss
@@ -1,0 +1,5 @@
+@media (max-width: 640px) and (min-width: 320px) {
+  .wrapper {
+    max-width: 98%;
+  }
+}
