+++ src/contentscript/overlay/root/components/Overlay/adaptive.scss
@@ -1,0 +1,40 @@
+@media (min-width: 320px) and (max-width: 640px) {
+  .overlay {
+    width: 100%;
+    min-width: 100%;
+  }
+
+  .wrapper {
+    overflow: hidden;
+
+    width: 100%;
+    height: 100%;
+    padding: 17px 15px 7px;
+  }
+
+  .inner {
+    flex: 1 1 auto;
+
+    max-width: 100%;
+    height: 100%;
+    padding: 37px 0 0;
+  }
+
+  .header {
+    margin-bottom: 35px;
+  }
+
+  .children {
+    height: 80%;
+  }
+
+  .toolbar {
+    position: relative;
+    top: 0;
+    left: 0;
+  }
+
+  .searchBlock {
+    top: inherit;
+  }
+}
