+++ src/contentscript/overlay/root/components/OverlayTab/adaptive.scss
@@ -1,0 +1,50 @@
+@import "../../assets/styles/color.variables.scss";
+
+@media (min-width: 320px) and (max-width: 640px) {
+  .tab {
+ 
+    display: flex;
+    justify-content: space-between;
+
+    width: max-content;
+    max-width: none;
+    height: 52px;
+    margin: 0;
+
+    border-radius: 10px;
+    background: #fff;
+  }
+
+  .list {
+    display: flex;
+    align-items: center;
+    flex-direction: row;
+
+    padding-left: 10px;
+  }
+
+  .item {
+    margin-top: 0;
+    padding: 0 10px;
+
+    &:not(:last-child) {
+      margin: 0;
+    }
+
+    &:nth-child(2) {
+      margin: 0;
+    }
+  }
+
+  .isOpenWallet {
+    &::before {
+      border-radius: 10px;
+      background-color: rgb(255 255 255 / 0%);
+    }
+  }
+
+  .tabNotActive {
+    width: auto;
+    max-width: min-content;
+  }
+}
